refactor: change saldo column type from double to bigdecimal

+++ src/main/java/com/luiz/devops/dtos/conta/ContaResponseDto.java
@@ -2,11 +2,12 @@ package com.luiz.devops.dtos.conta;
 
 import com.luiz.devops.dtos.movimentacoes.MovimentacoesResponseDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record ContaResponseDto(
         UUID id,
         List<MovimentacoesResponseDto> movimentacoes,
-        double saldo) {
+        BigDecimal saldo) {
 }
