+++ src/main/java/com/luiz/devops/dtos/movimentacoes/MovimentacoesRequestDto.java
@@ -5,11 +5,12 @@ import com.luiz.devops.enums.ValueOfEnum;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record MovimentacoesRequestDto(
         @NotNull UUID contaId,
-        @NotNull @Positive double valor,
+        @NotNull @Positive BigDecimal valor,
         @ValueOfEnum(enumClass = OperacaoEnum.class) String tipoMovimentacao
 ) {
 }
