+++ src/main/java/com/luiz/devops/dtos/movimentacoes/MovimentacoesResponseDto.java
@@ -1,9 +1,10 @@
 package com.luiz.devops.dtos.movimentacoes;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record MovimentacoesResponseDto(
         long id,
-        double valor,
+        BigDecimal valor,
         LocalDateTime data
 ) {}
